# docs/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Payment service",
    "description": "Service for managing payments and reservations",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "PaymentService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/payment/compensate": {
      "post": {
        "summary": "Компенсация денег (отмена резервирования)",
        "operationId": "PaymentService_CompensatePayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentserviceCompensatePaymentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentserviceCompensatePaymentRequest"
            }
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    },
    "/api/v1/payment/reserve": {
      "post": {
        "summary": "Резервирование денег",
        "operationId": "PaymentService_ReservePayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentserviceReservePaymentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentserviceReservePaymentRequest"
            }
          }
        ],
        "tags": [
          "PaymentService"
        ]
      }
    }
  },
  "definitions": {
    "paymentserviceCompensatePaymentRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "Сообщение для запроса компенсации денег"
    },
    "paymentserviceCompensatePaymentResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/paymentserviceResponseStatus"
        }
      },
      "title": "Сообщение для ответа на компенсацию денег"
    },
    "paymentserviceReservePaymentRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "Сообщение для запроса резервирования денег"
    },
    "paymentserviceReservePaymentResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/paymentserviceResponseStatus"
        }
      },
      "title": "Сообщение для ответа на резервирование денег"
    },
    "paymentserviceResponseStatus": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "SUCCESS",
        "INSUFFICIENT_QUANTITY",
        "INTERNAL_ERROR"
      ],
      "default": "UNKNOWN",
      "title": "Общий enum для статусов ответов"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}